Add frontend build script
