+++ frontend/mouse.ts
@@ -1,50 +1,38 @@
-import { sum } from "./sum.js";
+export function handleMouseMove(
+  event: MouseEvent,
+  xPos: HTMLInputElement,
+  yPos: HTMLInputElement
+) {
+  // On a MouseEvent, draw a dot where the cursor is if it is inside the listening element.
+  //
+  // Write the x and y coordinates to xPos and yPos.
+  let dot, pageX, pageY;
 
-window.onload = () => {
-  console.log(sum(5, 4));
-  const xPos = document.querySelector("#xpos") as HTMLInputElement | null;
-  const yPos = document.querySelector("#ypos") as HTMLInputElement | null;
+  event = event || window.event; // IE-ism
 
-  if (yPos == null || xPos == null) {
-    console.log("Could not find one or more of xPos and yPos");
-  }
-
-  const box = document.querySelector(".etch-a-sketch") as HTMLDivElement | null;
-
-  if (box == null) {
-    console.log("Could not find etch-a-sketch element");
+  // If pageX/Y aren't available and clientX/Y
+  // are, calculate pageX/Y - logic taken from jQuery
+  if (event.pageX != null && event.pageY != null) {
+    pageX = event.pageX;
+    pageY = event.pageY;
+  } else {
     return;
   }
-  box.onmousemove = handleMouseMove;
 
-  function handleMouseMove(event: MouseEvent) {
-    let dot, pageX, pageY;
+  // Add a dot to follow the cursor
+  dot = document.createElement("div");
+  dot.className = "dot";
+  dot.style.left = pageX + "px";
+  dot.style.top = pageY + "px";
+  document.body.appendChild(dot);
 
-    event = event || window.event; // IE-ism
-
-    // If pageX/Y aren't available and clientX/Y
-    // are, calculate pageX/Y - logic taken from jQuery
-    if (event.pageX != null && event.pageY != null) {
-      pageX = event.pageX;
-      pageY = event.pageY;
-    } else {
-      return;
-    }
-
-    // Add a dot to follow the cursor
-    dot = document.createElement("div");
-    dot.className = "dot";
-    dot.style.left = pageX + "px";
-    dot.style.top = pageY + "px";
-    document.body.appendChild(dot);
-
-    if (box != null) {
-      const boundingRect = box.getBoundingClientRect();
-      if (xPos != null && yPos != null) {
-        // Write the positions
-        xPos.value = (boundingRect.x - pageX).toString();
-        yPos.value = (boundingRect.y - pageY).toString();
-      }
+  if (event.currentTarget != null) {
+    const currentTarget = event.currentTarget as HTMLElement;
+    const boundingRect = currentTarget.getBoundingClientRect();
+    if (xPos != null && yPos != null) {
+      // Write the positions
+      xPos.value = (boundingRect.x - pageX).toString();
+      yPos.value = (boundingRect.y - pageY).toString();
     }
   }
-};
+}
