+++ frontend/package.json
@@ -5,7 +5,8 @@
   "devDependencies": {
     "@types/jest": "^29.0.2",
     "jest": "^29.0.3",
-    "ts-jest": "^29.0.1"
+    "ts-jest": "^29.0.1",
+    "tsc-watch": "^5.0.3"
   },
   "scripts": {
     "test": "jest"
