+++ frontend/sum.ts
@@ -1,3 +1,4 @@
+import { add } from "./add";
 export function sum(a: number, b: number) {
-  return a + b;
+  return add(a, b);
 }
