+++ sentrybot/static/js/add.js
@@ -1,0 +1,3 @@
+export function add(a, b) {
+    return a + b;
+}
