+++ sentrybot/static/js/sum.js
@@ -1,3 +1,4 @@
+import { add } from "./add.js";
 export function sum(a, b) {
-    return a + b;
+    return add(a, b);
 }
