+++ sentrybot/templates/base.html
@@ -3,7 +3,6 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <script>const exports = {};</script>
     {% block head %}
     {% endblock %}
 </head>
