+++ sentrybot/templates/face_detection.html
@@ -8,8 +8,7 @@
     position: absolute;
   }
 </style>
-<script type="text/javascript" src="static/js/hello.js"></script>
-<script type="module" src="static/js/mouse.js"></script>
+<script type="module" src="static/js/face_detection.js"></script>
 {% endblock %}
 {% block content %}
 <p>Testing, testing</p>
